LayoutTests/fast/js/dfg-arguments-osr-exit-multiple-blocks-before-exit-expected.txt

Tests aliased uses of 'arguments' that require reification of the Arguments object on OSR exit, in the case that there is some interesting control flow prior to the exit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS successfullyParsed is true

TEST COMPLETE
